requirements.txt
asgiref==3.3.1
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
flake8==3.8.4
flake8-django==1.1.1
gunicorn==20.0.4
importlib-metadata==3.3.0
isort==5.6.4
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
